feat: improve ssr html mismatch validation
