+++ .changeset/hungry-singers-share.md
@@ -1,0 +1,5 @@
+---
+"svelte": patch
+---
+
+feat: improve ssr html mismatch validation
