+++ packages/svelte/tests/runtime-runes/samples/invalid-html-ssr/Component.svelte
@@ -1,0 +1,1 @@
+<h1>foo</h1>
