+++ packages/svelte/tests/runtime-runes/samples/invalid-html-ssr/main.svelte
@@ -1,0 +1,7 @@
+<script>
+	import Component from "./Component.svelte";
+</script>
+
+<p>
+	<Component />
+</p>
